Move RaceViewerScreen to "race_viewer" file
