+++ lib/race_browser_screen/race_viewer.dart
@@ -1,8 +1,27 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:marklin_bluetooth/firebase/race.dart';
 import 'package:marklin_bluetooth/firebase/race_reference.dart';
 import 'package:marklin_bluetooth/race_browser_screen/car_viewer.dart';
+import 'package:marklin_bluetooth/utils.dart';
 import 'package:marklin_bluetooth/widgets.dart';
+
+/// Widget for displaying lap times and other information about [raceSnap].
+/// TODO: Add button for deleting race
+class RaceViewerScreen extends StatelessWidget {
+  const RaceViewerScreen({Key? key, required this.raceSnap}) : super(key: key);
+
+  final DocumentSnapshot<Race> raceSnap;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        appBar: AppBar(
+          title: Text("Viewing race: ${raceString(raceSnap)}"),
+        ),
+        body: RaceViewer(raceRef: RaceReference(docRef: raceSnap.reference)));
+  }
+}
 
 class RaceViewer extends StatelessWidget {
   const RaceViewer({Key? key, required this.raceRef}) : super(key: key);
