+++ lib/utils.dart
@@ -1,3 +1,6 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:marklin_bluetooth/firebase/race.dart';
+
 String dateString(
   DateTime date, {
   bool includeSecond = false,
@@ -12,4 +15,10 @@ String dateString(
   }
 
   return "$time $separator $day";
+}
+
+String raceString(DocumentSnapshot<Race> raceSnap) {
+  Race race = raceSnap.data()!;
+  DateTime date = race.date.toDate();
+  return (raceSnap.id == "current") ? "Current" : dateString(date);
 }
